Day 1/requirements.txt
streamlit
PyMuPDF
langchain
openai
faiss-cpu
